move all the scripts into a single file
